requirements.txt
black==24.10.0
click==8.1.7
mypy-extensions==1.0.0
packaging==24.2
paho-mqtt==2.1.0
pathspec==0.12.1
platformdirs==4.3.6
tomli==2.1.0
typing_extensions==4.12.2
